build pose-detector as a library
